Factor out js into a separate file
